adding changes

+++ src/app/chocolate-list/chocolate-list.component.ts
@@ -1,8 +1,8 @@
-import { Component, OnInit} from '@angular/core';
+import { Component, EventEmitter, OnInit, Output} from '@angular/core';
 import { ChocolateDataService } from '../chocolates-data-service.service';
 import { Chocolate } from './chocolate';
 import { Subscription } from 'rxjs';
-import { FormChocolatesComponent } from "../form-chocolates/form-chocolates.component";
+
 
 @Component({
   selector: 'app-chocolate-list',
@@ -12,13 +12,16 @@ import { FormChocolatesComponent } from "../form-chocolates/form-chocolates.comp
 export class ChocolateListComponent implements OnInit {
   subscription: Subscription | any;
   chocolates: Chocolate[] = [];
-  model: Chocolate = { id: 0, image: '', name: '', brand: '', price: 0, stock: 0};
-  form : FormChocolatesComponent
+
+  data: Chocolate = { id: 0, image: '', name: '', brand: '', price: 0, stock: 0};
+
+  @Output()
+  deleted: EventEmitter<string> = new EventEmitter<string>();
+
   constructor( private Chocolatedataservice : ChocolateDataService) {}
 
   ngOnInit(): void {
-
-    /*cuando se carga la pagina*/
+    /*when the page load*/
     this.Chocolatedataservice.getAll()
     .subscribe(chocolates => this.chocolates = chocolates );
     this.subscription = this.Chocolatedataservice.refresh$.subscribe(() => (
@@ -27,18 +30,28 @@ export class ChocolateListComponent implements OnInit {
   }
 
   getAll() {
+
     this.Chocolatedataservice.getAll()
     .subscribe(chocolates => this.chocolates = chocolates);
 
   }
   onsubmit() {
-    this.Chocolatedataservice.post(this.model)
+
+    this.Chocolatedataservice.post(this.data)
     .subscribe((response: Chocolate) => console.log(response));
 
   }
+
   deletechocolate(chocolate : Chocolate) {
-    this.Chocolatedataservice.delete(chocolate.id)
-    .subscribe((response: Chocolate) => console.log(response));
+
+    if (this.data){
+      this.Chocolatedataservice.delete(chocolate.id)
+     .subscribe((response: Chocolate) => console.log(response));
+    }
+    else {
+      this.deleted.emit("Se ha eliminado con exito")
+    }
+
   }
   }
